refactor: remove pkg folder from uibackend server
