+++ uibackend/server/dashboard_findings_trends.go
