+++ uibackend/server/gen.go
@@ -15,4 +15,4 @@
 
 package server
 
-//go:generate go run github.com/deepmap/oapi-codegen/v2/cmd/oapi-codegen --config=../server.cfg.yaml ../../openapi.yaml
+//go:generate go run github.com/deepmap/oapi-codegen/v2/cmd/oapi-codegen --config=server.cfg.yaml ../openapi.yaml
